+++ NEWS.md
@@ -3,6 +3,7 @@
 - add option `plot` to `TheilSen`. `FALSE` can be useful when analysing data to extract the trend components and plot in otherways and when the `TheilSen` plot is not required.
 - add option `silent` to `TheilSen` to avoid printing updates to trend fitting. By default it is `FALSE`.
 - fix wrong ordering of names in `timeVariation` when more than one pollutant
+- fix date parsing issues in `selectByDate`.
 
 # openair 2.2-3